Extract fixture content
